+++ TossClone/ImageLiteral.swift
@@ -1,0 +1,12 @@
+//
+//  ImageLiteral.swift
+//  TossClone
+//
+//  Created by 리아 on 2022/07/03.
+//
+
+import Foundation
+
+enum ImageLiteral {
+    static let logo = "toss"
+}
